+++ webpack.config.js
@@ -4,11 +4,13 @@ var karma = require('karma');
 var PORT = 8000;
 open('http://localhost:' + PORT);
 
-karma.server.start({
+var server = new karma.Server({
   configFile: __dirname + '/karma.conf.js',
   autoWatch: true,
   singleRun: false
 });
+
+server.start();
 
 module.exports = {
   entry: __dirname + '/src/angular-bootstrap-confirm.js',