+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.4.1 (2015-10-02)
+* Allow handle-focus to be set to false if set as an attribute.
+
+# 0.4.0 (2015-09-28)
+* Auto focus the confirm button when opening the popover. Thanks to @andreptb! 
+
+# 0.3.1 (2015-08-29)
+* Make `is-open` optional for angular 1.3 compatibility
+
 # 0.3.0 (2015-07-28)
 * Add is-open property to control the visibility of the popover